feat(retrace): resolve pct masks per-image in ApplyMasks
Add Pct field to pixel.Rect and change X/Y/Width/Height from int to float64
to support percentage-based mask specifications. Introduce ResolveRects to
convert Pct masks to absolute pixels per image dimensions, called by
ApplyMasks for each image independently so two screenshots of different sizes
sharing the same flow-keyed masks each resolve correctly to their own
dimensions.

+++ retrace/diff/pixel/trim.go
@@ -53,7 +53,7 @@ func TrimUniformBorder(img *image.RGBA) (cropped *image.RGBA, kept Rect, ok bool
 		return nil, Rect{}, false // too small to carry any comparison signal
 	}
 
-	kept = Rect{X: left - b.Min.X, Y: top - b.Min.Y, Width: kw, Height: kh}
+	kept = Rect{X: float64(left - b.Min.X), Y: float64(top - b.Min.Y), Width: float64(kw), Height: float64(kh)}
 	if kept.X == 0 && kept.Y == 0 && kw == w && kh == h {
 		return img, kept, true // already tight: no border found, left untouched
 	}
